Drop the 'terminfo' dependency
Refs #3696; this should make it easier to get the compiler installed,
because without the terminfo dependency, there's no need to dynamically
link to libtinfo. This removes what is probably the most common
stumbling block in getting the compiler installed and functioning
correctly. The cost of this is that the repl is not quite as good at
editing long input lines; see e.g.
https://github.com/purescript/purescript/issues/3696#issuecomment-657282303